+++ app/components/FooterCopyrights.tsx
@@ -1,7 +1,7 @@
 export function FooterCopyrights() {
   return (
     <div className="container pt-6 pb-11">
-      <p className="text-center text-base/[24px] tracking-[-0.35px]">
+      <p className="text-center text-sm/[20px] tracking-[-0.3px]">
         © 2025 Rahmat Hidayatullah — didesain dan dikembangkan dengan 💚 di
         Singaparna, Tasikmalaya.
       </p>