+++ app/components/NotionRenderer/common/Paragraph.tsx
@@ -4,36 +4,40 @@ import Image from "next/image";
 import { RichText } from "./RichText";
 
 export function Paragraph({ block }: { block: ParagraphBlockObjectResponse }) {
-  const regex = /<LocalImage\s+src="([^"]+)"\s+alt="([^"]+)"\s*\/>/;
+  const regex =
+    /<LocalImage\s+src="([^"]+)"\s+alt="([^"]+)"(?:\s+maxWidth="([^"]+)")?\s*\/>/;
   const matchAsImage = block.paragraph.rich_text[0]?.plain_text.match(regex);
 
   if (matchAsImage) {
     const src = matchAsImage[1];
     const alt = matchAsImage[2];
+    const maxWidth = matchAsImage[3];
 
     if (src.endsWith(".gif")) {
       return (
         // eslint-disable-next-line @next/next/no-img-element
         <img
-          width={1200}
-          height={720}
+          width={680}
+          height={382}
           src={src}
           alt={alt}
           loading="lazy"
           className="mb-7 rounded-xl overflow-hidden"
+          style={{ maxWidth }}
         />
       );
     }
     return (
       <Image
-        width={1200}
-        height={720}
+        width={680}
+        height={382}
         src={src}
         alt={alt}
         priority={false}
         loading="lazy"
         quality={90}
-        className="mb-7 rounded-xl overflow-hidden"
+        className="mb-7 rounded-xl overflow-hidden mx-auto"
+        style={{ maxWidth }}
       />
     );
   }