+++ app/components/NotionRenderer/index.tsx
@@ -2,11 +2,13 @@ import { ListBlockChildrenResponse } from "@notionhq/client";
 
 import { Blockquote } from "./common/Blockquote";
 import { BulletedListItem } from "./common/BulletedListItem";
+import { CodeBlock } from "./common/CodeBlock";
 import { Heading1 } from "./common/Heading1";
 import { Heading2 } from "./common/Heading2";
 import { Heading3 } from "./common/Heading3";
 import { isBlockObjectResponse } from "./utils";
 import { Paragraph } from "./common/Paragraph";
+import { Callout } from "./common/Callout";
 
 type NotionRendererProps = {
   listBlockChildren?: ListBlockChildrenResponse | null;
@@ -42,6 +44,12 @@ export function NotionRenderer({ listBlockChildren }: NotionRendererProps) {
                   index={index}
                 />
               );
+
+            case "code":
+              return <CodeBlock key={block.id} block={block} />;
+
+            case "callout":
+              return <Callout key={block.id} block={block} />;
 
             default:
               return null;