+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### ???
+### Added
 
--
+- Support articles filtering based on the running environment. New `NEXT_PUBLIC_APP_ENV` env needed.
+- Added necessary image assets for the "Mengenal Git dan GitHub" blog.
+- Added support for Notion block renderer of type "code" with `NotionRenderer/common/CodeBlock.tsx` component.
+
+### Fixed
+
+- Rendered titles and descriptions correctly in SelectedBlogSection.tsx and on the article detail page.
 
 ## [v0.5.2] - 2025-10-25