+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support articles filtering based on the running environment. New `NEXT_PUBLIC_APP_ENV` env needed.
 - Added necessary image assets for the "Mengenal Git dan GitHub" blog.
 - Added support for Notion block renderer of type "code" with `NotionRenderer/common/CodeBlock.tsx` component.
+- Support `maxWidth` attribute for <LocalImage> tag in notion.
+- Local images for blog with title: "Mengenal Git dan GitHub."
 
 ### Changed